Add SVG output
Added a function to generate a basic SVG version of a generated maze with some basic options to configure the output.

I am tempted to add an image/png output, but that would seem to add to many extra dependencies to this create, so is probably best done externally. SVG is just text, so much less overhead.